refactor(ai-providers): isolate account posture action
